backend/requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
drf-spectacular==0.26.0
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1 